Backend/requirements.txt
aiosmtplib==4.0.1
annotated-types==0.7.0
anyio==4.2.0
bcrypt==3.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.110.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.27.0
idna==3.7
openai==1.6.1
passlib==1.7.4
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.13
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
